fix(ui): keep agent tool policy out of the initial bundle
